+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # Rabbit Sky Client Changelog
 
+## Version 0.1.5
+- Skip versioning so it's the same as server.
+- Improve bandwidth usage, using formatted message instead of CSV.
+- Embed Chat add background color.
+
 ## Version 0.1
 - First Release